Remove micro-ros-agent middleware support and use MAVROS only
This setup should not require FMU firmware changes like tuning the dds_topics.yaml file

Use custom fork of MAVROS to enable MAVLink gimbal protocol v2 support
