--- docker/docker-compose.gisnav-ardupilot.yaml
@@ -1,6 +1,0 @@
-version: "3.4"
-
-services:
-  # Runs Ardupilot MAVROS launch configuration instead of default PX4 micro-ros
-  gisnav:
-    command: ros2 launch gisnav ardupilot.launch.py
