+++ launch/params/autopilot_node.yaml
@@ -1,2 +1,2 @@
-ardupilot_node:
+autopilot_node:
   ros__parameters: []
