+++ launch/params/mock_gps_node.yaml
@@ -1,2 +1,2 @@
-px4_node:
+mock_gps_node:
   ros__parameters: []
