--- launch/params/mock_gps_node_ardupilot.yaml
@@ -1,3 +1,0 @@
-mock_gps_node:
-  ros__parameters:
-    px4_micrortps: False  # Set to False for ArduPilot MAVROS
