--- launch/params/mock_gps_node_px4.yaml
@@ -1,3 +1,0 @@
-mock_gps_node:
-  ros__parameters:
-    px4_micrortps: True  # Set to False for ArduPilot MAVROS
